Optimised the header file

+++ volume/volume.h
@@ -24,9 +24,8 @@
 
 // REQUIRED SUBHEADERS
 #include "../cvect.h"
-#include "../section/section.h"
 
 //FUNCTION TO FIND THE VOLUME OF PARALLELOPIPED
-float parellopiped(int *x, int *y, int *z);
+int parellopiped(int *x, int *y, int *z);
 
 #endif
